+++ package.json
@@ -18,8 +18,8 @@
     "db:migrate": "drizzle-kit migrate"
   },
   "dependencies": {
-    "@auth/drizzle-adapter": "^1.2.0",
-    "@libsql/client": "^0.6.2",
+    "@auth/drizzle-adapter": "^1.1.0",
+    "@libsql/client": "^0.6.0",
     "@radix-ui/react-avatar": "^1.0.4",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-icons": "^1.3.0",
@@ -33,11 +33,11 @@
     "drizzle-orm": "^0.30.10",
     "libsql": "^0.3.18",
     "lucide-react": "^0.379.0",
-    "next": "14.2.3",
-    "next-auth": "beta",
+    "next": "14.2.4",
+    "next-auth": "5.0.0-beta.18",
     "next-themes": "^0.3.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "^18.3.0",
+    "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.1",
     "resend": "^3.2.0",
     "tailwind-merge": "^2.3.0",
@@ -46,21 +46,21 @@
   },
   "devDependencies": {
     "@types/bcryptjs": "^2.4.6",
-    "@types/node": "^20.13.0",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "autoprefixer": "^10.4.19",
     "drizzle-kit": "^0.21.4",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "14.2.3",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.4",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "husky": "^9.0.11",
-    "lint-staged": "^15.2.5",
-    "postcss": "^8.4.38",
-    "prettier": "^3.3.0",
-    "prettier-plugin-tailwindcss": "^0.5.14",
-    "tailwindcss": "^3.4.3",
+    "lint-staged": "^15.2.4",
+    "postcss": "^8",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.6.3",
+    "tailwindcss": "^3.4.4",
     "typescript": "^5.4.5"
   }
 }
