+++ src/app/(auth)/add-password/addPasswordForm.tsx
@@ -2,7 +2,7 @@
 // import { useActionState } from 'react';
 import { useState } from 'react';
 import { useRouter } from 'next/navigation';
-import { addPassword } from '@/actions/authAction';
+import { addPassword } from '@/actions/auth';
 import { useFormState } from 'react-dom';
 import { SubmitButton } from '@/components/SubmitButton';
 import { useEffect } from 'react';
