+++ src/app/(auth)/onboarding/OnBoardingForm.tsx
@@ -2,7 +2,7 @@
 // import { useActionState } from 'react';
 import { useState } from 'react';
 import { useRouter } from 'next/navigation';
-import { onBoarding } from '@/actions/authAction';
+import { onBoarding } from '@/actions/auth';
 import { useFormState } from 'react-dom';
 import { SubmitButton } from '@/components/SubmitButton';
 import { useEffect } from 'react';
