+++ src/app/(dashboard)/profile/change-password/changePasswordForm.tsx
@@ -2,7 +2,7 @@
 // import { useActionState } from 'react';
 import { useState } from 'react';
 import { useRouter } from 'next/navigation';
-import { changePassword } from '@/actions/authAction';
+import { changePassword } from '@/actions/auth';
 import { useFormState } from 'react-dom';
 import { SubmitButton } from '@/components/SubmitButton';
 import { useEffect } from 'react';
